requirements.txt
mcp>=1.2.0
httpx>=0.27.0
PyYAML>=6.0.1
uvicorn>=0.30.0
starlette>=0.37.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
oracledb>=2.0.0
